Refactored and Updated Code with comments.
