+++ backend/secure-chat-express/app/api/routes/routes.js
@@ -1,13 +1,10 @@
 const mongoose = require('mongoose');
-const crypto = require('crypto');
-const hash = crypto.createHash('sha1');
 const User = mongoose.model('User');
 const Conversation = mongoose.model('Conversation');
 const Message = mongoose.model('Message');
 const friendController = require('../controllers/friends');
 const conversationController = require('../controllers/conversation');
 const userController = require('../controllers/user');
-const words = require('../../wordJson/wordData');
 
 module.exports = function(app, passport) {
     // Route for logging in.
@@ -27,76 +24,12 @@ module.exports = function(app, passport) {
     );
 
     // Route for getting a user's publicKey
-    app.get('/api/v1/user/publicKey/:id',
-        passport.authenticate(['facebook-token']),
-        function(req, res) {
-            //  Check if the user was authenticated
-            if (req.user) {
-                var otherUserId = req.params.id;
-                if (!otherUserId) {
-                    res.status(400).json({'error': 'OtherUserID required in parameter.'});
-                    return;
-                }
-
-                User.findOne({'facebook.id': otherUserId}, function(err, user) {
-                    var publicKey = user.publicKey.keys.pgp;
-                    if (err) {
-                        res.status(500).json(err);
-                    } else if (!publicKey) {
-                        res.status(404).json({'error' : 'Public key not set for the queried user.'});
-                    } else {
-                        hash.update(publicKey);
-                        var digests = hash.digest('hex').toUpperCase().match(/.{1,2}/g);
-                        for (var i = 0; i < digests.length; i++) {
-                            digests[i] = words[digests[i]];
-                        }
-                        res.status(200).json(
-                            {
-                                'publicKey' : publicKey,
-                                'readableKey' : digests.join(" ")
-                            }
-                        );
-                    }
-                });
-            } else {
-                // Respond with Unauthorized access.
-                res.status(401).json({'error': 'Access Not Authorized.'});
-            }
-        }
-    );
+    app.get('/api/v1/user/publicKey/:id', passport.authenticate(['facebook-token']),
+        userController.getPublicKey);
 
     // Route for posting publicKey
-    app.put('/api/v1/user/publicKey',
-        passport.authenticate(['facebook-token']),
-        function(req, res) {
-            // Check if the user was authenticated
-            if (req.user) {
-                var thisUser = req.user;
-                var publicKey = req.body.publicKey;
-                if (!publicKey) {
-                    res.status(400).json({'error' : 'publicKey required in body.'});
-                    return;
-                }
-                hash.update(publicKey);
-                var digests = hash.digest('hex').toUpperCase().match(/.{1,2}/g);
-                for (var i = 0; i < digests.length; i++) {
-                    digests[i] = words[digests[i]];
-                }
-                thisUser.publicKey.keys.pgp = publicKey;
-                thisUser.publicKey.keys.readable = digests.join(" ");
-                thisUser.save(function(err, user) {
-                    if (err) {
-                        res.status(500).json(err);
-                    } else {
-                        res.status(201).json(user);
-                    }
-                });
-            } else {
-                // Respond with Unauthorized access.
-                res.status(401).json({'error': 'Access Not Authorized.'});
-            }
-        }
-    );
+    app.put('/api/v1/user/publicKey', passport.authenticate(['facebook-token']),
+        userController.putPublicKey);
 
     // Route for getting user information based on their facebook id.
     app.get('/api/v1/user/:id', passport.authenticate(['facebook-token']),
@@ -106,34 +39,9 @@ module.exports = function(app, passport) {
     app.get('/api/v1/user', passport.authenticate(['facebook-token']),
         userController.findUserByQuery);
 
+    // Route for getting the list of friends.
     app.get('/api/v1/user/friend/get', passport.authenticate(['facebook-token']),
-        function(req, res) {
-            if (req.user) {
-                if (req.user.friends.length <= 0) {
-                    res.status(404).json({'error' : 'No friends found for user'});
-                }
-                var friends = [];
-                var count = 0;
-                req.user.friends.forEach(function(friendId) {
-                    User.findOne({'facebook.id' : friendId}, function(err, friend) {
-                        if (err) {
-                            res.status(500).json(err);
-                        } else if (friend) {
-                            count++;
-                            friends.push(friend);
-                        }
-                    }).then(function() {
-                        if (count === req.user.friends.length) {
-                            res.status(200).json(friends);
-                        }
-                    });
-                });
-            } else {
-                // Respond with Unauthorized access.
-                res.status(401).json({'error': 'Access Not Authorized.'});
-            }
-        }
-    );
+        userController.getFriend);
 
     // Route for deleting a friend.
     app.put('/api/v1/user/friend/delete', passport.authenticate(['facebook-token']),
